+++ app/javascript/components/datasets/initial-state.js
@@ -40,6 +40,11 @@ export default {
       error: null,
       center: {}
     },
+    location: {
+      list: [],
+      selectedLocation: {},
+      area: {}
+    },
     areaOfInterest: {
       drawing: false,
       clearing: false,