+++ app/javascript/components/fsp-maps/actions.js
@@ -289,6 +289,36 @@ export const fetchNearbyArea = createThunkAction('ANALYSIS/fetchNearby', () => (
     });
 });
 
+// Analysis - Analyze Patterns
+export const setLocationSelected = createAction('ANALYSIS/setLocationSelected');
+export const setLocationsList = createAction('ANALYSIS/setLocationsList');
+export const setLocationArea = createAction('ANALYSIS/setLocationArea');
+export const fetchLocations = createThunkAction('ANALYSIS/fetchLocations', () => (dispatch, getState) => {
+  const { iso } = getState().fspMaps.common;
+  const locationsSql = encodeURIComponent(replace(JURISDICTIONS_SQL, { iso }));
+
+  fetch(`${window.FSP_CARTO_API}?q=${locationsSql}&api_key=${window.FSP_CARTO_API_KEY}`)
+    .then(response => response.ok && response.json())
+    .then((data) => {
+      dispatch(setLocationsList(data.rows));
+    });
+});
+
+export const fetchLocationArea = createThunkAction('ANALYSIS/fetchLocationArea', () => (dispatch, getState) => {
+  const { selectedLocation } = getState().fspMaps.analysis.location;
+  const { value: jurisdictionId } = selectedLocation;
+  const locationAreaSql = encodeURIComponent(
+    replace(JURISDICTION_GEOMETRY_SQL, { jurisdictionId })
+  );
+
+  fetch(`${window.FSP_CARTO_API}?q=${locationAreaSql}&api_key=${window.FSP_CARTO_API_KEY}`)
+    .then(response => response.ok && response.json())
+    .then((data) => {
+      dispatch(setLocationArea(JSON.parse(data.rows[0].st_asgeojson)));
+    });
+});
+
+
 // Analysis - area of interest
 export const setAreaOfInterest = createAction('ANALYSIS/setAreaOfInterest');
 export const setAreaOfInterestArea = createAction('ANALYSIS/setAreaOfInterestArea');